Added actual sleep record editing form
Form helpers require lots of ActiveRecord methods, so using a mock in
the view spec is extremely painful.  For that reason, we just generate a
full model object and assign it.  It kind of sucks that views can't be
tested in isolation if form helpers are used, but it makes a bit of
sense.  Form helpers are a rails feature and it makes sense for them to
depend on ActiveRecord models.  If we make the views depend on them,
then it's reasonable to expect them to also depend on ActiveRecord.  I
guess the way around that is to not use the form helper, which we may
want to do anyway because we're probably going to want some special
frontend affordances on this page to deal with sleep and wake times.